Move DownloadAir -> Main Project + Remove Unused
