--- LegendaryChat/MainWindow.xaml
@@ -1,8 +1,0 @@
-﻿<Window x:Class="LegendaryChat.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="MainWindow" Height="350" Width="525">
-    <Grid>
-        
-    </Grid>
-</Window>
