--- Patcher/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="SharpZipLib" version="0.86.0" targetFramework="net45" />
-</packages>
