--- RiotPatch/StuffThatYouShouldn'tHave/LatestVersion.cs
@@ -1,6 +1,0 @@
-﻿namespace RiotPatch.StuffThatYouShouldn_tHave
-{
-    public class LatestVersion
-    {
-    }
-}
