--- RiotPatch/Uncompress.cs
@@ -1,6 +1,0 @@
-﻿namespace RiotPatch
-{
-    public class Uncompress
-    {
-    }
-}
